+++ api/src/resources/mod.rs
@@ -1,5 +1,6 @@
 pub mod attachments;
 pub mod audit;
+pub mod auth;
 pub mod bucket;
 pub mod bucket_statistics;
 pub mod comment;